app/requirements.txt
PyQt6==6.7.1
pyqtgraph==0.13.7
paho-mqtt==2.1.0
requests==2.32.3
pandas==2.2.2
numpy==2.1.3
scikit-learn==1.5.2
reportlab==4.2.5
uvicorn==0.30.6
fastapi==0.115.0
# Optional
scipy==1.13.1
joblib==1.4.2
PyYAML==6.0.2
